refactor(ui): move bucket field popup components to molecules
Relocate popup-related components from atoms to molecules for clearer structure and organization.
Update import paths accordingly and improves inner field configuration handling for special field cases.
